+++ src/ds5/ds5-thermal-monitor.h
@@ -2,40 +2,11 @@
 // Copyright(c) 2021 Intel Corporation. All Rights Reserved.
 #pragma once
 
-#include "option.h"
 #include "sensor.h"
 #include "device-calibration.h"
 
 namespace librealsense
 {
-    template<typename T>
-    class protected_xu_option : public uvc_xu_option<T>
-    {
-    public:
-        protected_xu_option(uvc_sensor& ep, platform::extension_unit xu, uint8_t id, std::string description)
-            : uvc_xu_option<T>(ep, xu, id, description)
-        {}
-
-        protected_xu_option(uvc_sensor& ep, platform::extension_unit xu, uint8_t id, std::string description, const std::map<float, std::string>& description_per_value)
-            : uvc_xu_option<T>( ep, xu, id, description, description_per_value)
-        {}
-
-        void set(float value) override
-        {
-            std::lock_guard<std::mutex> lk(_mtx);
-            uvc_xu_option<T>::set(value);
-        }
-
-        float query() const override
-        {
-            std::lock_guard<std::mutex> lk(_mtx);
-            return uvc_xu_option<T>::query();
-        }
-
-    protected:
-
-        mutable std::mutex _mtx;
-    };
 
     class  ds5_thermal_monitor
     {
