+++ src/option.h
@@ -324,6 +324,7 @@ namespace librealsense
         std::function<void(const option&)> _record = [](const option&) {};
     };
 
+    // XU control with exclusing access to setter/getters
     template<typename T>
     class uvc_xu_option : public option
     {
@@ -404,6 +405,35 @@ namespace librealsense
         std::string         _desciption;
         std::function<void(const option&)> _recording_function = [](const option&) {};
         const std::map<float, std::string> _description_per_value;
+    };
+
+    template<typename T>
+    class protected_xu_option : public uvc_xu_option<T>
+    {
+    public:
+        protected_xu_option(uvc_sensor& ep, platform::extension_unit xu, uint8_t id, std::string description)
+            : uvc_xu_option<T>(ep, xu, id, description)
+        {}
+
+        protected_xu_option(uvc_sensor& ep, platform::extension_unit xu, uint8_t id, std::string description, const std::map<float, std::string>& description_per_value)
+            : uvc_xu_option<T>( ep, xu, id, description, description_per_value)
+        {}
+
+        void set(float value) override
+        {
+            std::lock_guard<std::mutex> lk(_mtx);
+            uvc_xu_option<T>::set(value);
+        }
+
+        float query() const override
+        {
+            std::lock_guard<std::mutex> lk(_mtx);
+            return uvc_xu_option<T>::query();
+        }
+
+    protected:
+
+        mutable std::mutex _mtx;
     };
 
     template<class T, class R, class W, class U>
